requirements.txt
langgraph>=0.1.9
litellm>=1.35.0
pydantic>=2.7
python-dotenv>=1.0.1
PyYAML>=6.0.1
httpx>=0.27.0
uvloop>=0.19.0; sys_platform != 'win32'
vllm